shift code of payments.tsx to billings.tsx
