+++ src/frontend/src/App.tsx
@@ -16,7 +16,6 @@ import PlanBilling from "./pages/PlanBilling";
 import Settings from "./pages/Settings";
 import OnboardingStep4 from "./pages/OnboardingStep4";
 import TempTester from "./modals/TempTester";
-import Payment from './pages/Payment'
 
 function App() {
     const action = useNavigationType();
@@ -124,7 +123,6 @@ function App() {
             <Route path="/channels" element={<ConnectChannels />} />
             <Route path="/billing" element={<PlanBilling />} />
             <Route path="/settings" element={<Settings />} />
-            <Route path="/payments" element={<Payment />} />
         </Routes>
     );
 }
